Error correcting at Cache Hit - when reading a cached file.
When I tested using curl - I tested using " curl -iS http://localhost:8080/http://http.badssl.com/" for the second time, to see if the cache-hit scenario works fine.
There was an error at Cache Hit - the server would print it is a cache hit, but the file was not sent to the client from the cache and the origin server was contacted.
I went through trial and error several times and solved the issue. The issues were: the file was not opened in the binary format, the use of readlines command and concatenating issue with string '>' and binary cacheData.

Hence, the cache file opening format was changed to 'rb', used 'read' command to read cacheFile and printed cacheData separately.