Tested the the ability to handle origin server ports that are specified in the URL.
Tested using http://portquiz.net:8080/ and it worked well.

However one error was detected - a cache hit occured when I tested with http://portquiz.net:666.

This means the cache location path did not include the port number and did not check for it.

Therefore the port number was added to cache location.